Adds more styled-system options to table and moves table to table.js file,
